model separated to main model and scene`s model, most of last controller code moved into view
