+++ settingsClone/settingsClone/Sources/Scenes/Detail/Controller/DetailController.swift
