+++ settingsClone/settingsClone/Sources/Scenes/Detail/Model/DetailModel.swift
@@ -1,0 +1,11 @@
+//
+//  DetailModel.swift
+//  settingsClone
+//
+//  Created by Nikita Alpatiev on 2/3/23.
+//
+
+struct DetalModel {
+    let image: String
+    let title: String
+}
