+++ settingsClone/settingsClone/Sources/Scenes/Detail/View/DetailView.swift
@@ -1,0 +1,14 @@
+//
+//  DetailView.swift
+//  settingsClone
+//
+//  Created by Nikita Alpatiev on 2/3/23.
+//
+
+import UIKit
+
+class DetailView: UIView {
+
+   
+
+}
