--- settingsClone/settingsClone/Sources/Scenes/Main/Model/SettingsModel.swift
@@ -1,81 +1,0 @@
-//
-//  SettingsModel.swift
-//  settingsClone
-//
-//  Created by Nikita Alpatiev on 1/1/23.
-//
-
-import Foundation
-
-// MARK: - Data model
-
-struct Cell: Codable {
-    let image: String
-    let title: String
-    
-    var subtitle: String? = nil
-    var badges: Int? = nil
-    var switcher: Bool? = nil
-}
-
-struct Section: Codable {
-    let name: String
-    var cells: [Cell]
-}
-
-struct Settings: Codable {
-    var sectionData = [Section]()
-}
-
-// MARK: - Model for first screen settings
-
-struct SettingsModel {
-    
-    static var resourceName = "settings"
-    static var list = [Section]()
-    
-    static func updateCell(from path: IndexPath?, cell: Cell) {
-        guard let indexPath = path else {
-            return
-        }
-        guard indexPath.section < list.count else {
-            return
-        }
-        guard indexPath.row < list[indexPath.section].cells.count else {
-            return
-        }
-        list[indexPath.section].cells[indexPath.row] = cell
-    }
-    
-    static func returnCellFrom(_ sectionIndex: Int, _ rowIndex: Int) -> Cell? {
-        guard sectionIndex < SettingsModel.list.count else {
-            return nil
-        }
-        guard rowIndex < SettingsModel.list[sectionIndex].cells.count else {
-            return nil
-        }
-        return SettingsModel.list[sectionIndex].cells[rowIndex]
-    }
-    
-    static func fetch() -> Settings? {
-        guard let settingsPath = Bundle.main.path(forResource: SettingsModel.resourceName, ofType: "json") else {
-            return nil
-        }
-        
-        let url = URL(fileURLWithPath: settingsPath)
-        
-        do {
-            let data = try Data(contentsOf: url)
-            return try JSONDecoder().decode(Settings.self, from: data)
-        } catch {
-            return nil
-        }
-    }
-    
-    static func setupData() {
-        if let result = SettingsModel.fetch() {
-            SettingsModel.list = result.sectionData
-        }
-    }
-}
-
