--- settingsClone/settingsClone/Sources/Scenes/Main/View/SettingsView.swift
@@ -1,20 +1,0 @@
-//
-//  SettingsView.swift
-//  settingsClone
-//
-//  Created by Nikita Alpatiev on 2/2/23.
-//
-
-import UIKit
-
-class SettingsView: UIView {
-
-    /*
-    // Only override draw() if you perform custom drawing.
-    // An empty implementation adversely affects performance during animation.
-    override func draw(_ rect: CGRect) {
-        // Drawing code
-    }
-    */
-
-}
